cland format check

+++ tsMuxer/lpcmStreamReader.cpp
@@ -15,8 +15,7 @@ namespace
 {
 uint32_t FOUR_CC(const char a, const char b, const char c, const char d)
 {
-    return my_ntohl(static_cast<uint32_t>(a) << 24 | static_cast<uint32_t>(b) << 16 | static_cast<uint32_t>(c) << 8 |
-                    d);
+    return my_ntohl(static_cast<uint32_t>(a) << 24 | b << 16 | c << 8 | d);
 }
 }  // namespace
